add random data

+++ app/src/main/java/org/lab/biometro/model/TempModel.java
@@ -3,6 +3,6 @@ package org.lab.biometro.model;
 public class TempModel {
 
     public String time = "15:00:00";
-    public int value = 100;
+    public double value = 100;
 
 }
